Updated package.json to remove deps and correct license
Growing much closer to releasing. Hopefully in the next couple of days; I just
don't want to rush it. Though, at the same time, I've been noticing projects
popping up with very similar / exact names to this one. A project named "ease"
was added to the npm repository and another "ease.js" is on GitHub, although
it's made no progress. As such, I want to ensure I reserve the name in npm.

I've been testing the new library and work and noticed only a couple minor
issues, primiarly due to misuse of the library. Looking good.