Beginning concepts for article on JavaScript freedom
I am currently working with RMS on a release announcement that touches on this
topic, but felt that a permenant article on the actual project would serve an
even greater purpose of attrating additional free software supporters.

This is by no means meant to replace RMS' ``The JavaScript Trap'' article; it
merely expands upon it.