Altered license templates for combined files with section 7 exception
As suggested by RMS in The JavaScript Trap:
  <https://gnu.org/philosophy/javascript-trap.html>

This does increase the size of the minified file a bit---the header is now
about 1kB of uncompressed text (which will hopefully comprses nicely with
the rest of the minified file). That said, ease.js will be continuing to
grow in size, bandwidth is becoming less and less important, and the license
is very important, especially in our goal to spread the philosophy of
software freedom.