Remove GitHub link from download page
The section is maintained, but provides a link to whyfreejs.html instead of
a link to the repository, to discourage its use.

At some point, I'll include the reply from GitHub stating that they have no
intent of liberating the client code, or making the website work properly
without JavaScript.