Add two hashes to signchk-ignore
I had just gotten back into easejs development and must not have had git
configured to automatically sign commits.

More notably, though, is that all of my old commits are signed with a
now-revoked signature.  I'll have to come up with a plan for that with
regards to signature verification; not dealing with that right now.

* tools/signchk-ignore: Add hashes of two commits I mistakenly did not sign.